requirements.txt
# Database connectivity
psycopg2-binary==2.9.9

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Logging
colorama==0.4.6

# Metrics
prometheus-client==0.19.0

